pde/hasegawa-wakatani/requirements.txt
numpy
cupy-cuda13x
matplotlib